requirements.txt
Flask==2.0.1
python-dotenv==0.20.0
boto3==1.18.21
gunicorn==20.1.0
Werkzeug==2.0.1
